requirements.txt
torch>=2.1.0
torch-geometric>=2.4.0
torch-scatter>=2.1.2
torch-sparse>=0.6.18
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
hdbscan>=0.8.33
matplotlib>=3.8.0
seaborn>=0.13.0
pyyaml>=6.0
